Write times as UTC rather than using TZ=timezone
If we were to write TZ=timezone, then strict conformance with the
RFC 5545 specification) requires that this timezone be fully specified
(including BST and GMT start dates and offsets, etc).
Of course, these dates are the last Sundays in March/October,
so they change which would mean recalculating the VTIMEZONE
specification each year. Yuck.