Improvements
Added window.onerror to catch all errors not otherwise handled.

renderTable guesses that a cell containing a 'large' numeric value
is an Excel serial date and renders it as such.
This allows users to enter dates as 1/2/2023 etc.
This functionality is duplicated, probably unecessarily, in parseDate.

generateICal allows dates to have a year other than the default year
but gives a warning.

Code needs cleaning up, eg to remove several log messages.