Dates and times
Tried to deal with dates and times
- Used cellStyles: true, cellDates: true to get XLSX.read to return Date objects
- Dealt this Steve Gray use of non-integers to represent a time
This seems to work provided the spreadsheet is VERY careful with formats.

However, this fails badly with test-calendar.xlsx with many cells appearing
as Nan:NaN. The cells had values including
  Saturday 1 April
  Saturday
  April
The reason seems to be that these cells had formats suggesting they were
dates. Clearly "Saturday" and "April" are not dates. "Saturday 1 April"
had a format "hh:mm".
It seem that we cannot rely on Excel format types but need to use the raw
values.